ENH:S'afegeix com a opció de configuració que l'usuari pugui escollir si esborrar els estudis vells passat un determinat número de dies i escollir si quan no hi hagi suficient espai per descarregar nou estudis s'han d'esborrar estudis vells per alliberar espai

+++ starviewer/src/inputoutput/queryscreen.cpp
@@ -202,7 +202,14 @@ void QueryScreen::setSeriesToSeriesListWidgetCache()
 
 void QueryScreen::deleteOldStudies()
 {
-    m_qdeleteOldStudiesThread.deleteOldStudies();
+    StarviewerSettings settings;
+
+    /*Mirem si està activada la opció de la configuració d'esborrar els estudis vells no visualitzats en un número de dies determinat
+      fem la comprovació, per evitar engegar el thread si no s'han d'esborrar els estudis vells*/
+    if (settings.getDeleteOldStudiesHasNotViewedInDays())
+    {
+        m_qdeleteOldStudiesThread.deleteOldStudies();
+    }
 }
 
 void QueryScreen::checkDatabaseImageIntegrity()
